svn merge -r4172:5369 https://nozomi/pro60/branches/tnt .
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@5384 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/products/rendergirl/test.sh
@@ -23,6 +23,11 @@ compareSceneImages() {
     return 0
 }
 
+if [ "$MAYA_SDK" == "" ] ; then
+    echo "MAYA_SDK not set - skipping tests"
+    exit 0
+fi
+
 rm -f TEST_IMAGES/*
 rm -f TESTMODELS/*.x60
